Module seax_svm::cellStable

SVM cell types.

A cell in the VM can be either an atom (single item, either unsigned int, signed int, float, or string), a pointer to a list cell, or an instruction.

Reexports

pub use self::SVMCell::*;

pub use self::Atom::*;

Enums

Atom
SVM atom types.

Inst
SVM instruction types

SVMCell