Enum seax_svm::IOEvent

pub enum IOEvent {
    Req,
    Buf(char),
}

Unstable

A VM state's IO action

Take note that this will eventually be replaced with memory-mapped IO when the main memory management scheme is done; therefore, it should never be marked as stable. Consider this struct and anything that depends on it to be an ugly hack.

Variants

Req
Unstable

A character was requested from the buffer

Buf
Unstable

A character was buffered