Module seax_svm::slistStable

Singly-linked list and stack implementations.

List<T> is a singly-linked cons list with boxed items. Stack<T> is defined as a trait providing stack operations(push(), pop(), and peek()), and an implementation for List.

Reexports

pub use slist::List::{Cons, Nil};

Structs

ListIterator
Wraps a List to allow it to be used as an Iterator

Enums

List
Singly-linked cons list.

Traits

Stack
Common functions for an immutable Stack abstract data type.