Enum seax_svm::slist::ListStable

pub enum List<T> {
    Cons(T, Box<List<T>>),
    Nil,
}

Singly-linked cons list.

This is used internally to represent list primitives in the machine.

Variants

Cons

Cons cell containing a T and a link to the tail

Nil

The empty list.

Methods

impl<T> List<T>

Public implementation for List.

fn new() -> List<T>

Creates a new empty list

fn prepend(self, it: T) -> List<T>

Prepends the given item to the list.

This is an O(1) operation.

Arguments

item - the item to prepend

Return Value

The list with the new head item prepended

Examples

let mut a_list: List<isize> = List::new();
a_list = a_list.prepend(1);
assert_eq!(a_list, list![1]);

a_list = a_list.prepend(2);
assert_eq!(a_list, list![2,1]);

fn append(&mut self, it: T)

Appends an item to the end of the list.

This is an O(n) operation.

Arguments

item - the item to append

Examples

let mut a_list: List<isize> = List::new();
a_list.append(1);
assert_eq!(a_list, list![1]);

a_list.append(2);
assert_eq!(a_list, list![1,2]);

fn length(&self) -> usize

Returns the length of the list.

Examples

let a_list = list!(1,2,3,4);
assert_eq!(a_list.length(), 4)

fn iter<'a>(&'a self) -> ListIterator<'a, T>

Provide a forward iterator

fn last(&self) -> &T

Returns the last element of the list

Examples

let a_list = list!(1,2,3,4);
assert_eq!(a_list.last(), &4)