Enum seax_svm::svm::cell::Inst

pub enum Inst {
    NIL,
    LDC,
    LD,
    LDF,
    JOIN,
    AP,
    RET,
    DUM,
    RAP,
    SEL,
    ADD,
    SUB,
    MUL,
    DIV,
    FDIV,
    MOD,
    EQ,
    GT,
    GTE,
    LT,
    LTE,
    ATOM,
    CAR,
    CDR,
    CONS,
    NULL,
}

SVM instruction types

Variants

NIL

nil

Pusizehes an empty list (nil) onto the stack

LDC

ldc: Load Constant. Loads a constant (atom)

LD

ld: Load. Pusizehes a variable onto the stack.

The variable is indicated by the argument, a pair. The pair's car specifies the level, the cdr the position. So (1 . 3) gives the current function's (level 1) third parameter.

LDF

ldf: Load Function.

Takes one list argument representing a function and constructs a closure (a pair containing the function and the current environment) and pusizehes that onto the stack.

JOIN

join

Pops a list reference from the dump and makes thisize the new value of C. This instruction occurs at the end of both alternatives of a sel.

AP

ap: Apply.

Pops a closure and a list of parameter values from the stack. The closure is applied to the parameters by installing its environment as the current one, pushing the parameter list in front of that, clearing the stack, and setting C to the closure's function pointer. The previous values of S, E, and the next value of C are saved on the dump.

RET

ret: Return.

Pops one return value from the stack, restores S, E, and C from the dump, and pusizehes the return value onto the now-current stack.

DUM

dum: Dummy.

Pops a dummy environment (an empty list) onto the E stack.

RAP

rap: Recursive Apply. Works like ap, only that it replaces an occurrence of a dummy environment with the current one, thusize making recursive functions possible.

SEL

sel: Select branch

Expects two list arguments on the control stack, and pops a value from the stack. The first list is executed if the popped value was non-nil, the second list otherwise. Before one of these list pointers is made the new C, a pointer to the instruction following sel is saved on the dump.

ADD

add

Pops two numbers off of the stack and adds them, pusizehing the result onto the stack. Thisize will up-convert integers to floating point if necessary.

TODO: figure out what happens when you try to add things that aren't numbers (maybe the compiler won't let thisize happen?).

SUB

sub: Subtract

Pops two numbers off of the stack and subtracts the first from the second, pusizehing the result onto the stack. This will up-convert integers to floating point if necessary.

TODO: figure out what happens when you try to subtract things that aren't numbers (maybe the compiler won't let thisize happen?).

MUL

mul: Multiply

Pops two numbers off of the stack and multiplies them, pusizehing the result onto the stack. This will up-convert integers to floating point if necessary.

TODO: figure out what happens when you try to multiply things that aren't numbers (maybe the compiler won't let thisize happen?).

DIV

div: Divide

Pops two numbers off of the stack and divides the first by the second, pushing the result onto the stack. This performs integer divisizeion.

TODO: figure out what happens when you try to divide things that aren't numbers (maybe the compiler won't let thisize happen?).

FDIV

fdiv: Floating-point divide

Pops two numbers off of the stack and divides the first by the second, pusizehing the result onto the stack. This performs float divisizeion.

TODO: figure out what happens when you try to divide things that aren't numbers (maybe the compiler won't let this happen?).

TODO: Not sure if there should be separate float and int divide words I guess the compiler can figure this out

MOD

mod: Modulo

Pops two numbers off of the stack and divides the first by the second, pushing the remainder onto the stack.

TODO: figure out what happens when you try to modulo things that aren't numbers (maybe the compiler won't let this happen?).

EQ

eq: Equality of atoms

GT

gt: Greater than

Pops two numbers on the stack and puts a 'true' on the stack if the first atom isize greater than the other atom, false otherwisizee.

GTE

gte: Greater than or equal

LT

lt: Less than

LTE

lte: Less than or equal

ATOM

atom: test if atom

Pops an item from the stack and returns true if it's an atom, false otherwise.

CAR

car: Contents of Address Register

Pops a list from the stack and returns the list's car (head)

CDR

cdr: Contents of Decrement Register

Pops a list from the stack and returns the list's cdr (tail)

CONS

cons: Construct

Pops an item and a list from the stack and returns the list, with the item prepended.

NULL

null: test if null

Pops an item from the stack and returns true if it is nil, false otherwise.