Enum seax_svm::svm::slist::List

pub enum List<T> {
    Cons(T, Box<List<T>>),
    Nil,
}

Singly-linked cons list.

This is used internally to represent list primitives in the machine.

Variants

Cons

Cons cell containing a T and a link to the tail

Nil

The empty list.

Methods

impl<T> List<T>

Public implementation for List.

fn new() -> List<T>

Creates a new empty list

fn prepend(self, it: T) -> List<T>

Prepends the given item to the list.

Returns the list containing the new head item. This is an O(1) operation.

fn append(self, it: T)

Appends an item to the end of the list.

This is an O(n) operation.

fn length(&self) -> usize

Returns the length of the list.

fn iter<'a>(&'a self) -> ListIterator<'a, T>

Provide a forward iterator

fn last(&self) -> &T